Applied some MatchModel improvements from paq8gen

+++ Shared.hpp
@@ -59,7 +59,7 @@ public:
 
       //MatchModel
       struct {
-        uint32_t length3;     //used by SSE stage and RecordModel
+        uint8_t length3;     //used by SSE stage and RecordModel
         uint8_t expectedByte; //used by SSE stage and RecordModel
       } Match{};
